WIP scoring. refactoring, permissions, tests

+++ forecast_app/views.py
@@ -3,7 +3,6 @@ import io
 import json
 import logging
 import time
-from itertools import groupby
 
 import boto3
 import django_rq
@@ -25,6 +24,7 @@ from django.views.generic import DetailView, ListView
 from forecast_app.forms import ProjectForm, ForecastModelForm
 from forecast_app.models import Project, ForecastModel, Forecast, TimeZero, ScoreValue, Score, ScoreLastUpdate
 from forecast_app.models.project import Target, Location
+from forecast_app.models.score import _write_csv_score_data_for_project
 from forecast_app.models.upload_file_job import UploadFileJob, upload_file_job_s3_file
 from forecast_repo.settings.base import S3_UPLOAD_BUCKET_NAME
 from utils.flusight import flusight_location_to_data_dict
@@ -97,6 +97,9 @@ def _rq_jobs_for_queue(rq_queue):
 
 
 def empty_rq(request):
+    if not is_user_ok_admin(request.user):
+        raise PermissionDenied
+
     queue = django_rq.get_queue()  # name='default'
     queue.empty()
     messages.success(request, "Emptied the queue.")
@@ -104,6 +107,9 @@ def empty_rq(request):
 
 
 def delete_upload_file_jobs(request):
+    if not is_user_ok_admin(request.user):
+        raise PermissionDenied
+
     # NB: delete() runs in current thread. recall pre_delete() signal deletes corresponding S3 object (the uploaded
     # file)
     UploadFileJob.objects.all().delete()
@@ -256,6 +262,9 @@ def _location_to_actual_max_val(loc_tz_date_to_actual_vals):
 #
 
 def delete_scores(request):
+    if not is_user_ok_admin(request.user):
+        raise PermissionDenied
+
     # NB: delete() runs in current thread
     Score.objects.all().delete()
     messages.success(request, "Deleted all Scores.")
@@ -263,6 +272,9 @@ def delete_scores(request):
 
 
 def delete_score_last_updates(request):
+    if not is_user_ok_admin(request.user):
+        raise PermissionDenied
+
     # NB: delete() runs in current thread
     ScoreLastUpdate.objects.all().delete()
     messages.success(request, "Deleted all ScoreLastUpdates.")
@@ -401,84 +413,6 @@ def download_scores(request, project_pk):
     writer = csv.writer(response)
     _write_csv_score_data_for_project(writer, project)
     return response
-
-
-def _write_csv_score_data_for_project(csv_writer, project):
-    """
-    Writes all ScoreValue data for project into csv_writer. There is one column per ScoreValue BUT: all Scores are on
-    one line. Thus, the row 'key' is the (fixed) first five columns:
-
-        `ForecastModel.name, TimeZero.timezero_date, season, Location.name, Target.name`
-
-    Followed on the same line by a variable number of ScoreValue.value columns, one for each Score. Score names are in
-    the header. An example header and first few rows:
-
-        model,           timezero,    season,    location,  target,          constant score,  Absolute Error
-        gam_lag1_tops3,  2017-04-23,  2017-2018  TH01,      1_biweek_ahead,  1                <blank>
-        gam_lag1_tops3,  2017-04-23,  2017-2018  TH01,      1_biweek_ahead,  <blank>          2
-        gam_lag1_tops3,  2017-04-23,  2017-2018  TH01,      2_biweek_ahead,  <blank>          1
-        gam_lag1_tops3,  2017-04-23,  2017-2018  TH01,      3_biweek_ahead,  <blank>          9
-        gam_lag1_tops3,  2017-04-23,  2017-2018  TH01,      4_biweek_ahead,  <blank>          6
-        gam_lag1_tops3,  2017-04-23,  2017-2018  TH01,      5_biweek_ahead,  <blank>          8
-        gam_lag1_tops3,  2017-04-23,  2017-2018  TH02,      1_biweek_ahead,  <blank>          6
-        gam_lag1_tops3,  2017-04-23,  2017-2018  TH02,      2_biweek_ahead,  <blank>          6
-        gam_lag1_tops3,  2017-04-23,  2017-2018  TH02,      3_biweek_ahead,  <blank>          37
-        gam_lag1_tops3,  2017-04-23,  2017-2018  TH02,      4_biweek_ahead,  <blank>          25
-        gam_lag1_tops3,  2017-04-23,  2017-2018  TH02,      5_biweek_ahead,  <blank>          62
-
-    Notes:
-    - `season` is each TimeZero's containing season_name, similar to Project.timezeros_in_season().
-    - we use get_valid_filename() to ensure values are CSV-compliant, i.e., no commas, returns, tabs, etc. Using that
-      function is as good as any :-)
-    - we use groupby to group row 'keys' so that all score values are together
-    """
-    # re: scores order: it is crucial that order matches query ORDER BY ... sv.score_id so that columns match values
-    scores = Score.objects.all().order_by('pk')
-
-    # write hearder
-    SCORE_CSV_HEADER = ['model', 'timezero', 'season', 'location', 'target'] + \
-                       [get_valid_filename(score.name) for score in scores]
-    csv_writer.writerow(SCORE_CSV_HEADER)
-
-    # get the raw rows - sorted so groupby() will work
-    sql = """
-        SELECT f.forecast_model_id, f.time_zero_id, sv.location_id, sv.target_id, sv.score_id, sv.value
-        FROM {scorevalue_table_name} AS sv
-               INNER JOIN {score_table_name} s ON sv.score_id = s.id
-               INNER JOIN {forecast_table_name} AS f ON sv.forecast_id = f.id
-               INNER JOIN {forecastmodel_table_name} AS fm ON f.forecast_model_id = fm.id
-        WHERE fm.project_id = %s
-        ORDER BY f.forecast_model_id, f.time_zero_id, sv.location_id, sv.target_id, sv.score_id;
-    """.format(scorevalue_table_name=ScoreValue._meta.db_table,
-               score_table_name=Score._meta.db_table,
-               forecast_table_name=Forecast._meta.db_table,
-               forecastmodel_table_name=ForecastModel._meta.db_table)
-    with connection.cursor() as cursor:
-        cursor.execute(sql, (project.pk,))
-        rows = cursor.fetchall()
-
-    # write grouped rows
-    forecast_model_id_to_obj = {forecast_model.pk: forecast_model for forecast_model in project.models.all()}
-    timezero_id_to_obj = {timezero.pk: timezero for timezero in project.timezeros.all()}
-    location_id_to_obj = {location.pk: location for location in project.locations.all()}
-    target_id_to_obj = {target.pk: target for target in project.targets.all()}
-    for (forecast_model_id, time_zero_id, location_id, target_id), score_id_value_grouper \
-            in groupby(rows, key=lambda _: (_[0], _[1], _[2], _[3])):
-        forecast_model = forecast_model_id_to_obj[forecast_model_id]
-        timezero = timezero_id_to_obj[time_zero_id]
-        location = location_id_to_obj[location_id]
-        target = target_id_to_obj[target_id]
-        season = '?'  # todo xx timezero.containing_season_name()
-        # ex score_groups: [(1, 18, 1, 1, 1, 1.0), (1, 18, 1, 1, 2, 2.0)]  # multiple scores per group
-        #                  [(1, 18, 1, 2, 2, 0.0)]                         # single score
-        score_groups = list(score_id_value_grouper)
-        # NB: if a score is missing then we need to use '' for it so that scores align with the header:
-        score_id_to_value = {score_group[-2]: score_group[-1] for score_group in score_groups}
-        score_values = [score_id_to_value[score.id] if score.id in score_id_to_value else None for score in scores]
-        csv_writer.writerow(
-            [get_valid_filename(forecast_model.name), timezero.timezero_date, get_valid_filename(season),
-             get_valid_filename(location.name), get_valid_filename(target.name)]
-            + score_values)
 
 
 #
